Improve delete security

+++ src/actions/DeleteAction.php
@@ -9,7 +9,7 @@
 namespace eseperio\filescatalog\actions;
 
 
-use app\helpers\ArrayHelper;
+use yii\helpers\ArrayHelper;
 use eseperio\admintheme\helpers\Html;
 use eseperio\filescatalog\controllers\DefaultController;
 use eseperio\filescatalog\dictionaries\InodeTypes;
@@ -20,6 +20,7 @@ use Yii;
 use yii\base\Action;
 use yii\web\BadRequestHttpException;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 
 class DeleteAction extends Action
 {
@@ -33,22 +34,46 @@ class DeleteAction extends Action
     {
         $model = $this->controller->findModel(Yii::$app->request->get('uuid'), File::class);
 
-
+        if ($model->isRoot())
+            throw new ForbiddenHttpException(Yii::t('filescatalog', 'Root node can not be deleted'));
         $parentUuid = $model->getParent()->select('uuid')->scalar();
 
         $rcvdHash = Yii::$app->request->post($this->module->secureHashParamName);
 
+
         if (!empty($rcvdHash) && $rcvdHash === $model->deleteHash && AclHelper::canDelete($model)) {
-            if (Yii::$app->request->post('delall')) {
-                if ($model->type === InodeTypes::TYPE_VERSION) {
-                    File::deleteAll([
-                        'id' => ArrayHelper::getColumn($model->original->versions, 'id')
+            if ($model->type === InodeTypes::TYPE_DIR) {
+                if (Yii::$app->request->post('confirm_text') === $model->getDeletionConfirmText()) {
+                    //Delete files one per each
+                    $descendantFiles = $model->getDescendants()->andWhere([
+                        'type' => [InodeTypes::TYPE_VERSION, InodeTypes::TYPE_FILE]
                     ]);
-                    $model->original->delete();
+                    foreach ($descendantFiles->batch(50) as $rows) {
+                        foreach ($rows as $row) {
+                            /* @var $row File */
+                            $row->delete();
+                        }
+                    }
+
+                    $model->deleteWithChildren();
+                } else {
+                    return $this->controller->render('delete', [
+                        'model' => $model
+                    ]);
                 }
             } else {
-                $model->delete();
+                if (Yii::$app->request->post('delall')) {
+                    if ($model->type === InodeTypes::TYPE_VERSION) {
+                        File::deleteAll([
+                            'id' => ArrayHelper::getColumn($model->original->versions, 'id')
+                        ]);
+                        $model->original->delete();
+                    }
+                } else {
+                    $model->delete();
+                }
             }
+
 
             return $this->controller->redirect(['index', 'uuid' => $parentUuid]);
         } else {
